fixed int conversion

+++ src/it/unimi/dsi/webgraph/labelling/ScatteredLabelledArcsASCIIGraph.java
@@ -149,7 +149,7 @@ public class ScatteredLabelledArcsASCIIGraph extends ImmutableSequentialGraph {
 	/**
 	 * The default label mapping function.
 	 */
-	public static final LabelMapping DEFAULT_LABEL_MAPPING = (prototype, representation) -> ((GammaCodedLongLabel) prototype).value = Integer.parseInt(new String(representation));
+	public static final LabelMapping DEFAULT_LABEL_MAPPING = (prototype, representation) -> ((GammaCodedLongLabel) prototype).value = Long.parseLong(new String(representation));
 
 	private static final Logger LOGGER = LoggerFactory.getLogger(ScatteredLabelledArcsASCIIGraph.class);
 	private final static boolean DEBUG = false;
